Rename Point3D to WorldPoint
